Delete all zeroes signature-check
Removes checking for all zeroes within the signature in order to proceed or skip the sms sending. The SMS provider is specified within the environment variables.
To confirm the OTP, the server checks for a valid provider. If no valid provider is specified within the environment, will behave as any token is valid.